% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Calculte API for Know Temperature Data}
\usage{
api(
  sp_df,
  ar_var = NULL,
  use_year = FALSE,
  flag_sp = FALSE,
  flag_month = FALSE,
  flag_ar = FALSE
)
}
\arguments{
\item{sp_df}{Input data. Must be a dataframe containing a column titled
"Binomial" which contains genus species binomial ("Poecile atricapillus"),
a column title "Ar" which has temperature values in degrees Celsius, and a
column titled "Month" that contains a numeric value from 1-12 or month codes
(ex. "Jan"). If Month of observed aridity recording is unknown users may
choose to include "Year" in the Month column which will have the function
calculate API based on Yearly Average API limits. NAs will be ignored.}

\item{ar_var}{Input string. Defaults to NULL which informs function to
calculate API for column titled "Ar". Options: "minmax", "all".
"minmax" indicates user has two columns with aridity data, one should be
titled "Amin", the other "Amax". "all" indicates user has three columns with
aridity data, should be titled "Amin", "Amax", and "Ar" ("Ar" indicating
mean aridity). Output columns for Amin = API_min, Amax = API_max,
Ar = API_mean when ar_var = "all".}

\item{use_year}{Boolean parameter, defualt FALSE. set to TURE to use yearly
aridity tolerance limits instead of Monthly limits.}

\item{flag_sp}{Boolean parameter, default FALSE. Set to true to add column
to final data frame that contains TRUE values for submitted species Binomials
that match API data set, and FALSE for rows that contain Binomials that do
not match those in the API data set. Used to Identify erroneous Binomials}

\item{flag_month}{Boolean parameter, default FALSE. Set to true to add column
to final data frame that contains TRUE values for appropriately submitted
Month formats and FALSE for inappropriately formatted Month values. Used to
Identify erroneous month submissions.}

\item{flag_ar}{Boolean parameter, default FALSE. Set to true to add
column(s) to final data frame that contains FALSE values if aridity
values are submitted as NA. TRUE is reported when no aridity errors are
submitted. Aids in identifying errors of aridity submission.}
}
\value{
A Dataframe.
}
\description{
\code{api()} Takes dataframe submitted by user and calculates Aridity Position
Index. Uses species Binomial and Month to extract aridity niche limits then
takes user submitted aridity values and calculates API using the aridity
niche limits add a column to submitted dataframe with API values.Aridity
Index values cannot be calculated for snow cover, and due to some species
ranges being completely covered in snow for some seasons, certain species
will have values of NA returned for API calculations.These NA values indicate
that the period of calculation is during a period with no determined
aridity index measurement for the species range.
}
\examples{
#example df:
species_data <- <- data.frame(
               Binomial = c("Poecile atricapillus", "Poecile atricapillus"),
               Month = c(1,2),
               AMax = c(5,5),
               AMin = c(2,1),
               Ar = c(0,3)
               )

#would return API for column Tm
results <- api(species_data)

#would return API for AMin and AMax
results <- api(species_data, ar_var = "minmax")

#would return API for AMin, AMax, and Ar
results <- api(species_data, ar_var = "all", use_year = TRUE,
                flag_sp = TRUE, flag_month = TRUE, flag_ar = TRUE)

}
