% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_check.R
\name{binomial_check}
\alias{binomial_check}
\title{Checks for matching species Binomial name}
\usage{
binomial_check(sp_binomial)
}
\arguments{
\item{sp_binomial}{Input string or vector. Must be a character vector}
}
\value{
A a data frame.
}
\description{
\code{binomial_check()} takes user submitted species binomial or list of binomials
and compares to the species names recorded in the TPI and API datafiles.
Returns True if there is a matching name and False if not. First letter of
Genus needs to be capitalized and species should be all lower case.
}
\examples{
#single species
binomial_check("Poecile atricapillus")
#multiple species
binomial_check(c("Poecile atricapillus","Melospiza melodia"))

sp_data <- c("Poecile atricapillus","Melospiza melodia")
binomial_check(sp_data)
}
