% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_niche.R
\name{generate_niche}
\alias{generate_niche}
\title{Creates niche limits from user specified inputs}
\usage{
generate_niche(
  sp_dist,
  niche_vars = NULL,
  precip = FALSE,
  time_period = NULL,
  monthly = TRUE,
  tmp_units = NULL
)
}
\arguments{
\item{sp_dist}{Input SpatVector, List of SpatVectors or Dataframe.
Input SpatVector(s) will be assumed to be from a single species. Can be
points or polygons. Dataframe must have 3 columns with the headings: "Group",
"lon", "lat". The "Group" column can include any numeric or character string
used to group observations together. "lon" must have longitude coordinates
and "lat" must have latitude coordinates.}

\item{niche_vars}{Input character string. Default NULL generates Thermal and
aridity index niche limits. Can be specified as "temp" which generates only
thermal niche limits, or "arid" which generates only Aridity index niche
limits. Temperature, precipitation and potential evapotransipration data are
extracted from Terra Climate v1.1 at a resolution of 2.5 minutes.}

\item{precip}{Boolean (TRUE/FALSE). When set to true precipitation niche
limits are generated instead of aridity index. Default FALSE. PMin, PMax
refer to lower and upper precipitation limits.}

\item{time_period}{Numeric Vector. Users can specify the years over which
thermal and aridity niche limits are calculated. Default NULL calculates
niche limits over the years 1961-1975. Time periods are limited to the years
1950-2025}

\item{monthly}{Boolean (TRUE/FALSE). Default TRUE calculates niche limits on
a monthly basis. FALSE calculates yearly average niche limits.}

\item{tmp_units}{Character String. Default thermal niche limits are calculated
in degrees Celsius. Options: "Fahrenheit", "Kelvin", convert thermal niche
limits to the specified temperature unit.}
}
\value{
A a data frame.
}
\description{
\code{generate_niche()} takes user submitted species distribution data and to
generate realized niche limits following the methods used in Watson & Kerr
(2025) Scientific Data (https://doi.org/10.1038/s41597-025-06130-1) When
all parameters are left default. Users can additionally specify time periods
for calculating niche limits. Since historical WorldClim monthly weather
data is not available to directly call in R we use historical monthly TMax
and TMin from TerraClimate. This will result in slightly different niche
limit estimates compared to those contained within this package.
}
\examples{
df <- data.frame(Group = c(1,1,1,1,1,1),
lat = c(41.334287,41.334287,41.334287,41.334287,41.334287,41.334287),
lon = c(-77.942020,-77.942020,-77.942020,-77.942020,-77.942020,-77.942020),

niche_lim <- generate_niche(df)

}
