% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_vars.R
\name{get_env_vars}
\alias{get_env_vars}
\title{Extracts Environmental Variables from Coordinates and Dates}
\usage{
get_env_vars(point_data, env_vars = "both", date_format = "monthly")
}
\arguments{
\item{point_data}{Input data. Must be a dataframe containing a column titled
"Lon" and "Lat", containing Longitude and Latitude in degrees. Depending on
date of observation users should include 3 separate columns for dates: "Year"
"Month" and "Day". All values for dates should be numeric. If interested in
getting seasonal average conditions data must include a column titled Season
and contain "spring", "summer", "fall", or "winter". Function accounts for
geographic location when considering season based on Hemisphere of
observation}

\item{env_vars}{Character string or vector. Data will be extracted for each
specified parameter. Options: "both", "tmp", "aridity". Mean monthly
temperature is not available through Terra Climate. If users want to use
mean monthly temperatures they need to acquire their data separately.}

\item{date_format}{Input character string to specify date resolution.
Defaults to "monthly" and uses Year and Month to extract data. Specify
"daily" if your date observations are contain Year,Month, and Day AND has
coordinates contained within the continental United States Years (1979-
present). Specify "seasonal" to get monthly average variable values for the
given season (generates 3 rows per input). Specify "yearly" to extract monthly
average variable values for each month (generates 12 rows per input).}
}
\value{
A Dataframe.
}
\description{
\code{get_env_vars} Takes dataframe submitted by user and extracts envrionmental
variable from Terra Climate. Users can specify Temperature, Aridirty, or
both. Can extract data based on Year, Year and Month, Year and Month and Day
or Season. Requires climateR, terra, and sf packages.
}
\examples{

df <- data.frame(Year = c(2000,2000,2000,2000,2000,2000),
Month = c(6,6,6,6,6,6),
Day = c(3,5,7,9,11,13),
Lat = c(41.334287,41.334287,41.334287,41.334287,41.334287,41.334287),
Lon = c(-77.942020,-77.942020,-77.942020,-77.942020,-77.942020,-77.942020),
Season = c("spring","spring","spring","spring","spring","spring"))

test <- get_env_vars(df, env_vars="aridity", date_format = "daily")

test <- get_env_vars(df, env_vars="both", date_format="monthly")

test <- get_env_vars(df, env_vars="tmp", date_format="yearly")

test <- get_env_vars(df, env_vars="both", date_format="seasonal")

}
