% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highertaxa_limits.R
\name{highertaxa_limits}
\alias{highertaxa_limits}
\title{Extracts Thermal and Aridity Tolerance Limits for Taxonomic Groups}
\usage{
highertaxa_limits(
  taxa_code,
  taxa_list,
  niche_limit = NULL,
  month_list = NULL,
  yr_avg = FALSE
)
}
\arguments{
\item{taxa_code}{Taxonomic level. Enter: "class","order","family",or "genus".}

\item{taxa_list}{Input string or vector of rank names. Must be a character
vector. All submissions should have a capitalized first letter.}

\item{niche_limit}{Input string. Specify "tpi" for thermal limits, "api" for
aridity limits, or "both" for both limits. Default is "both".}

\item{month_list}{Input number or character vector. Specify the corresponding
number for month of desired (1=Jan), or month code (Jan).
must be a number or character vector. Defaults to returning all months.}

\item{yr_avg}{Boolean paramater. Defaults to FALSE. if specified as TRUE
an additional row is included in the dataframe containing average yearly
niche limits.}
}
\value{
A dataframe containing the niche limit variables specified for the
taxanomic groups submitted.
}
\description{
\code{highertaxa_limits()} takes user submitted taxanomic group or list of groups
and extracts thermal and aridity position limits for each month. Optional
parameters can be specified to refine extracted limits to specific niche
limit, month, and yearly average limit values.
}
\examples{

highertaxa_limits(taxa_code = "class", taxa_list="Aves")

highertaxa_limits(taxa_code = "order",
taxa_list=c("Galliformes","Anseriformes"))

#specify TPI only
highertaxa_limits(taxa_code = "class", taxa_list="Aves",
niche_limit = "tpi")

#Specify January
highertaxa_limits(taxa_code = "class", taxa_list="Aves",
niche_limit = "tpi", month_list=1)

highertaxa_limits(taxa_code = "class", taxa_list="Aves",
niche_limit = "tpi", month_list="Jan")

#Get yearly average TPI values
highertaxa_limits(taxa_code = "class", taxa_list="Aves",
niche_limit = "tpi", month_list="Jan", yr_avg=TRUE)
}
