% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_limits.R
\name{species_limits}
\alias{species_limits}
\title{Extracts Thermal and Aridity Tolerance Limits}
\usage{
species_limits(
  sp_binomial,
  niche_limit = NULL,
  month_list = NULL,
  yr_avg = FALSE
)
}
\arguments{
\item{sp_binomial}{Input string or vector. Must be a character vector.}

\item{niche_limit}{Input string. Specify "tpi" for thermal limits, "api" for
aridity limits, or "both" for both limits. Default is "both".}

\item{month_list}{Input number or character vector. Specify the corresponding
number for month of desired (1=Jan), or month code (Jan).
must be a number or character vector. Defaults to returning all months.}

\item{yr_avg}{Boolean paramater. Defaults to FALSE. if specified as TRUE
an additional row is included in the dataframe containing average yearly
niche limits.}
}
\value{
A dataframe containing the niche limit variables specified for the
species submitted.
}
\description{
\code{species_limits()} takes user submitted species binomial or list of binomials
and extracts thermal and aridity position limits for each month. Optional
parameters can be specified to refine extracted limits to specific niche
limit, month, and yearly average limit values.
}
\examples{

species_limits("Poecile atricapillus")

species_limits(c("Poecile atricapillus","Melospiza melodia"))

sp_list <- c("Poecile atricapillus","Melospiza melodia")
species_limits(sp_list)

#specify TPI only
species_limits("Poecile atricapillus", niche_limit = "tpi")

#Specify January
species_limits("Poecile atricapillus", niche_limit = "tpi", month_list=1)
species_limits("Poecile atricapillus", niche_limit = "tpi", month_list="Jan")

Get yearly average TPI values
species_limits("Poecile atricapillus", niche_limit = "tpi", month_list=1, yr_avg=TRUE)
}
