% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list.R
\name{species_list}
\alias{species_list}
\title{Returns a list of species Binomial names from the niche limits datasets}
\usage{
species_list(class = NULL)
}
\arguments{
\item{class}{Input string. Options include: "mammal", "amphibian", "bird",
"reptile"}
}
\value{
A a data frame.
}
\description{
\code{species_list()} if left as NULL returns a data frame with columns for:
taxonomic Class, Order, Family, Genus, Species, and Binomial for all groups
of terrestrial vertebrates included in the niche limits data set. Users can
specify taxonomic class for sub sampling.
}
\examples{
#all species
all_list <- species_list()
#specific taxonomic class
mammal_list <- species_list("mammal")

}
