% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_check.R
\name{syn_check}
\alias{syn_check}
\title{Converts submitted species names to match IUCN and BirdLife names}
\usage{
syn_check(binomial_list)
}
\arguments{
\item{binomial_list}{Input string or vector of genus species binomials. Must
be a character vector. All submissions should have a capitalized first letter
of the genus name followed by a space and lower case species name: eg.
Melospiza melodia.}
}
\value{
A dataframe containing the user submitted name and matching IUCN or
BirdLife genus species binomial
}
\description{
\code{syn_check()} takes user submitted genus species binomial names and
creates a list that converts all submitted names to the appropriate matching
binomial utilized by the IUCN and BirdLife. Binomials are compared against
ITIS binomial synonyms. The list of binomial synonyms is not exhaustive and
any non matching names may need to be verified through other taxonomic
databases
}
\examples{

sp_list <- c("Melospiza melodia", "Zonotrichia albicollis")

Compared_df <- syn_check(sp_list)

sp_df <- data.frame(
               Binomial = c("Melospiza melodia", "Zonotrichia albicollis"),
               Month = c(1,2),
               TMin = c(0,1),
               TMax = c(15,13)
               )
df_Binomial_check <- syn_check(sp_df$Binomial)

#or

df_Binomial_check <- syn_check(sp_df[,"Binomial"])

}
