% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi.R
\name{tpi}
\alias{tpi}
\title{Calculte TPI for Know Temperature Data}
\usage{
tpi(
  sp_df,
  tmp_var = NULL,
  use_year = FALSE,
  flag_sp = FALSE,
  flag_month = FALSE,
  flag_tmp = FALSE
)
}
\arguments{
\item{sp_df}{Input data. Must be a dataframe containing a column titled
"Binomial" which contains genus species binomial ("Poecile atricapillus"),
a column title "Tm" which has temperature values in degrees Celcius, and a
column titled "Month" that contains a numeric value from 1-12 or month codes
(ex. "Jan"). If Month of observed temperature recording is unknown users may
choose to include "Year" in the Month column which will have the function
calculate TPI based on Yearly Average TPI limits. NAs will be ignored.}

\item{tmp_var}{Input string. Defaults to NULL which informs function to
calculate TPI for column titled "Tm". Options: "minmax", "all".
"minmax" indicates user has two columns with temperature data, one should be
titled "Tmin", the other "Tmax". "all" indicates user has three columns with
temperature data, should be titled "Tmin", "Tmax", and "Tm" ("Tm" indicating
mean temperature). Output columns for Tmin = TPI_min, Tmax = TPI_max,
Tm = TPI_mean when tmp_var = "all".}

\item{use_year}{Boolean parameter, defualt FALSE. set to TURE to use yearly
thermal tolerance limits when Month is NA}

\item{flag_sp}{Boolean parameter, default FALSE. Set to true to add column
to final data frame that contains TRUE values for submitted species Binomials
that match TPI data set, and FALSE for rows that contain Binomials that do
not match those in the TPI data set. Used to Identify erroneous Binomials}

\item{flag_month}{Boolean parameter, default FALSE. Set to true to add column
to final data frame that contains TRUE values for appropriately submitted
Month formats and FALSE for inappropriately formatted Month values. Used to
Identify erroneous month submissions.}

\item{flag_tmp}{Boolean parameter, default FALSE. Set to true to add
column(s) to final data frame that contains FALSE values if temperature
values are submitted as NA. TRUE is reported when no temperature errors are
submitted. Aids in identifying errors of temperature submission.}
}
\value{
A Dataframe.
}
\description{
\code{tpi()} Takes dataframe submitted by user and calculates Thermal Position
Index. Uses species Binomial and Date to extract thermal niche limits and
add a column to submitted dataframe with TPI values.
}
\examples{
#example df:
species_data <- data.frame(
               Binomial = c("Poecile atricapillus", "Poecile atricapillus"),
               Month = c(1,2),
               TMax = c(5,10),
               TMin = c(-2,-1),
               Tm = c(0,3)
               )

#would return TPI for column Tm
results <- tpi(species_data)

#would return TPI for TMin and TMax
results <- tpi(species_data, tmp_var = "minmax")

#would return TPI for TMin, TMax, and Tm
results <- tpi(species_data, tmp_var = "all", use_year = TRUE,
                flag_sp = TRUE, flag_month = TRUE, flag_tmp = TRUE)

}
